REFACTOR: Edit layout and fix profile page bug
Edited layout to show the camera related function on the right side so
that the users can see all the necessary info well. Also, fixed the
profile page bug. there was a bug that doesn't change the user's current
profile when navigating to the profile page from some other users'
profile page right away. By putting 'id' in the useEffect's [], it could
be solved.